// frontend/src/api/generated/models/constraintCatalogEntry.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * 途排智策 API
 * 途排智策前后端分离 MVP API
 * OpenAPI spec version: 0.1.0
 */
import type { ConstraintScopeField } from './constraintScopeField';
import type { ConstraintCatalogEntryHardnessItem } from './constraintCatalogEntryHardnessItem';
import type { ConstraintCatalogEntrySolverPaths } from './constraintCatalogEntrySolverPaths';
import type { ConstraintCatalogEntrySoftWeightHint } from './constraintCatalogEntrySoftWeightHint';
import type { ConstraintCatalogEntryAliasOf } from './constraintCatalogEntryAliasOf';

export interface ConstraintCatalogEntry {
  type: string;
  label: string;
  description?: string;
  scope?: string[];
  scope_fields?: ConstraintScopeField[];
  hardness: ConstraintCatalogEntryHardnessItem[];
  solver_paths?: ConstraintCatalogEntrySolverPaths;
  soft_weight_hint?: ConstraintCatalogEntrySoftWeightHint;
  alias_of?: ConstraintCatalogEntryAliasOf;
}